ui fix

+++ mandelbrot/AllSettings.h
@@ -1,10 +1,12 @@
 #ifndef ALLSETTINGS_H
 #define ALLSETTINGS_H
 
+#include <memory>
+
 struct AllSettings
 {
     int pow_deg = 2;
-    int max_step = 200;
+    size_t max_step = 200;
 };
 
 #endif // ALLSETTINGS_H
